+++ Changelog.md
@@ -1,3 +1,12 @@
+## 開發日誌
+
+# v0.1.3
+  - 可以重新加載 maintain 模組
+  - 實作 download-file, delete-file 功能
+  - 實作 list-child-process, stop-child-process 功能
+  - help 訊息簡化
+  - 修正重啟 bug
+
 # v0.1.2
   - 新增儲存綁定 id 的 json
   - query 指令未輸入 id 時, 查詢自己